// src/models/GlobalModel.ts
export interface GlobalModel {
    showFormPeriode: boolean;
    showFormKelas: boolean;
    showFormMapel: boolean;
    showFormSekolah: boolean;
    showFormSiswaKelas: boolean;
    showFormSiswa: boolean;
    showFormUser: boolean;
    showFormEditPeriode: any[];
    showFormEditKelas: any[];
    showFormEditMapel: any[];
    showFormEditSekolah: any[];
    showFormEditSiswaKelas: any[];
    showFormEditSiswa: any[];
    showFormEditUser: any[];
}